+++ test/test_sample.py
@@ -22,7 +22,7 @@ except ImportError:
 from cmdstanpy import _TMPDIR
 from cmdstanpy.cmdstan_args import CmdStanArgs, Method, SamplerArgs
 from cmdstanpy.model import CmdStanModel
-from cmdstanpy.stanfit import from_csv, CmdStanMCMC, RunSet
+from cmdstanpy.stanfit import CmdStanMCMC, RunSet, from_csv
 from cmdstanpy.utils import EXTENSION, cmdstan_version_at
 
 HERE = os.path.dirname(os.path.abspath(__file__))
@@ -52,7 +52,9 @@ class SampleTest(unittest.TestCase):
         for root, _, files in os.walk(DATAFILES_PATH):
             for filename in files:
                 _, ext = os.path.splitext(filename)
-                if ext.lower() in ('.o', '.hpp', '.exe', ''):
+                if ext.lower() in ('.o', '.d', '.hpp', '.exe', '') and (
+                    filename != ".gitignore"
+                ):
                     filepath = os.path.join(root, filename)
                     os.remove(filepath)
 
@@ -1540,6 +1542,46 @@ class CmdStanMCMCTest(unittest.TestCase):
             self.assertTrue(os.path.exists(profile_file))
             diagnostics_file = profile_fit.runset.diagnostic_files[i]
             self.assertTrue(os.path.exists(diagnostics_file))
+
+    def test_xarray_draws(self):
+        stan = os.path.join(DATAFILES_PATH, 'bernoulli.stan')
+        jdata = os.path.join(DATAFILES_PATH, 'bernoulli.data.json')
+        bern_model = CmdStanModel(stan_file=stan)
+        bern_fit = bern_model.sample(
+            data=jdata, chains=2, seed=12345, iter_warmup=100, iter_sampling=100
+        )
+        xr_data = bern_fit.draws_xr()
+        self.assertEqual(xr_data.theta.dims, ('chain', 'draw'))
+        self.assertTrue(
+            np.allclose(
+                xr_data.theta.transpose('draw', ...).values,
+                bern_fit.draws()[:, :, -1],
+            )
+        )
+        self.assertEqual(xr_data.theta.values.shape, (2, 100))
+
+        # test inc_warmup
+        bern_fit = bern_model.sample(
+            data=jdata,
+            chains=2,
+            seed=12345,
+            iter_warmup=100,
+            iter_sampling=100,
+            save_warmup=True,
+        )
+        xr_data = bern_fit.draws_xr(inc_warmup=True)
+        self.assertEqual(xr_data.theta.values.shape, (2, 200))
+
+        # test that array[1] and chains=1 are properly handled dimension-wise
+        stan = os.path.join(DATAFILES_PATH, 'bernoulli_array.stan')
+        jdata = os.path.join(DATAFILES_PATH, 'bernoulli.data.json')
+        bern_model = CmdStanModel(stan_file=stan)
+        bern_fit = bern_model.sample(
+            data=jdata, chains=1, seed=12345, iter_warmup=100, iter_sampling=100
+        )
+        xr_data = bern_fit.draws_xr()
+        self.assertEqual(xr_data.theta.dims, ('chain', 'draw', 'theta_dim_0'))
+        self.assertEqual(xr_data.theta.values.shape, (1, 100, 1))
 
 
 if __name__ == '__main__':
